+++ app/src/test/java/woowacourse/shopping/domain/RecentProductsTest.kt
@@ -4,31 +4,6 @@ import java.util.UUID
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 
-class RecentProducts(private val products: List<Product>) {
-    fun addProduct(newProduct: Product): RecentProducts {
-        var newProducts = products
-        if (newProducts.size >= 10) {
-            newProducts = newProducts.take(9)
-        }
-
-        val duplicate = newProducts.firstOrNull {
-            it.id == newProduct.id
-        }
-        if (duplicate != null) {
-            newProducts = products.filter { it.id != newProduct.id }
-        }
-        return RecentProducts(listOf(newProduct) + newProducts)
-    }
-
-    fun mostRecentProduct(): Product {
-        return products.first()
-    }
-
-    fun sizeOf(): Int {
-        return products.size
-    }
-}
-
 class RecentProductsTest {
     @Test
     fun `상품을 추가하면 가장 최신 위치에 들어간다`() {
